BaseSets Class

The BaseSets class implements a container to describe distinct objects that make up sets, along with element metadata and set metadata.

BaseSets(relations = DataFrame(element = character(0), set =
  character(0)), elementData, setData)

Arguments

relations
DataFrame. Two columns provide mapping relationships between "element" and "set".

elementData
DataFrame. Provide metadata for each unique element in relations$element.

setData
DataFrame. Provide metadata for each unique element in relations$set.

Value

A BaseSets object.

Slots

relations

DataFrame. Two columns provide mapping relationships between "element" and "set".

elementData

DataFrame. Provide metadata for each unique element in relations$element.

setData

DataFrame. Provide metadata for each unique element in relations$set.

See also

Examples

# Constructor ----

# Visually intuitive definition of sets
sets <- list(
  set1=c("A", "B"),
  set2=c("B", "C", "D"),
  set3=c("E"))

bs <- as(sets, "BaseSets")

# Coercing ----

# to list (gene sets)
ls1 <- as(bs, "list")
# to matrix (logical membership)
m1 <- as(bs, "matrix")

# Accessors ----

relations(bs)

#> DataFrame with 6 rows and 2 columns
#>      element        set
#>   <IdVector> <IdVector>
#> 1          A       set1
#> 2          B       set1
#> 3          B       set2
#> 4          C       set2
#> 5          D       set2
#> 6          E       set3

elementData(bs)

#> IdVector of length 5 with 5 unique identifiers
#> Ids: A, B, C, D, ...
#> Metadata:  (0 columns)

setData(bs)

#> IdVector of length 3 with 3 unique identifiers
#> Ids: set1, set2, set3
#> Metadata:  (0 columns)

# Dimensions ----

length(bs)

#> [1] 6

nElements(bs)

#> [1] 5

nSets(bs)

#> [1] 3

elementIds(bs)

#> [1] "A" "B" "C" "D" "E"

setIds(bs)

#> [1] "set1" "set2" "set3"

setLengths(bs)

#> set1 set2 set3 
#>    2    3    1 

elementLengths(bs)

#> A B C D E 
#> 1 2 1 1 1 

# Getters/Setters ----

bs1 <- bs
elementIds(bs1) <- paste0("gene", seq_len(nElements(bs)))
setIds(bs1) <- paste0("geneset", seq_len(nSets(bs)))