GOHits Class

The GOHits class extends the Hits class to represent hits that also describe relations between genes and sets using the Gene Ontology controlled vocabulary.

GOEvidenceCodes

GOOntologyCodes

GOHits(from = integer(0), to = integer(0),
  evidence = factor(character(0), names(GOEvidenceCodes)),
  ontology = factor(character(0), names(GOOntologyCodes)), nLnode = 0L,
  nRnode = 0L, ...)

Arguments

from, to
Two integer vectors of the same length. The values in from must be >= 1 and <= nLnode. The values in to must be >= 1 and <= nRnode.

evidence
factor. Levels must be values in names(GOEvidenceCodes).

ontology
factor. Levels must be values in names(GOOntologyCodes).

nLnode, nRnode
Number of left and right nodes.

...
Arguments metadata columns to set on the GOHits object. All the metadata columns must be vector-like objects of the same length as from, to, and membership.

Format

A named vector of length 26.

names

Code.

carat

Description.

...

Source

Evidence codes: http://geneontology.org/docs/guide-go-evidence-codes/

Namespaces: http://geneontology.org/docs/guide-go-evidence-codes/

Value

A GOHits object.

Details

This class does not define any additional slot to the Hits class. However, this class defines additional validity checks to ensure that every relation stored in a GOHits are respect the Gene Ontology evidence and ontology codes. Refer to GOOntologyCodes and GOEvidenceCodes for valid code and vocabulary.

Controlled vocabulary

Gene Ontology evidence codes were obtained from http://geneontology.org/docs/guide-go-evidence-codes/

Gene Ontology namespaces were obtained from http://geneontology.org/docs/ontology-documentation/

See also

Examples

# Controlled vocabulary ----

GOEvidenceCodes

#>                                                     EXP 
#>                              "Inferred from Experiment" 
#>                                                     IDA 
#>                            "Inferred from Direct Assay" 
#>                                                     IPI 
#>                    "Inferred from Physical Interaction" 
#>                                                     IMP 
#>                        "Inferred from Mutant Phenotype" 
#>                                                     IGI 
#>                     "Inferred from Genetic Interaction" 
#>                                                     IEP 
#>                      "Inferred from Expression Pattern" 
#>                                                     HTP 
#>              "Inferred from High Throughput Experiment" 
#>                                                     HDA 
#>            "Inferred from High Throughput Direct Assay" 
#>                                                     HMP 
#>        "Inferred from High Throughput Mutant Phenotype" 
#>                                                     HGI 
#>     "Inferred from High Throughput Genetic Interaction" 
#>                                                     HEP 
#>      "Inferred from High Throughput Expression Pattern" 
#>                                                     IBA 
#>   "Inferred from Biological characteristic of Ancestor" 
#>                                                     IBD 
#> "Inferred from Biological characteristic of Descendant" 
#>                                                     IKR 
#>          "Inferred loss due to absence of Key Residues" 
#>                                                     IRD 
#>                  "Inferred loss after Rapid Divergence" 
#>                                                     ISS 
#>       "Inferred from Sequence or structural Similarity" 
#>                                                     ISO 
#>                      "Inferred from Sequence Orthology" 
#>                                                     ISA 
#>                      "Inferred from Sequence Alignment" 
#>                                                     ISM 
#>                          "Inferred from Sequence Model" 
#>                                                     IGC 
#>                         "Inferred from Genomic Context" 
#>                                                     RCA 
#>         "Inferred from Reviewed Computational Analysis" 
#>                                                     TAS 
#>                            "Traceable Author Statement" 
#>                                                     NAS 
#>                        "Non-traceable Author Statement" 
#>                                                      IC 
#>                                   "Inferred by Curator" 
#>                                                      ND 
#>                          "No biological Data available" 
#>                                                     IEA 
#>                   "Inferred from Electronic Annotation" 

GOOntologyCodes

#>                   BP                   MF                   CC 
#> "Biological Process" "Molecular Function" "Cellular Component" 

# Constructor ----

from <- c(5, 2, 3, 3, 3, 2)
to <- c(11, 15, 5, 4, 5, 11)
ontology <- factor(c("BP", "BP", "BP", "MF", "MF", "CC"))
evidence <- factor(c("IEA", "IDA", "IEA", "IDA", "IEA", "IDA"))

gh <- GOHits(from, to, evidence, ontology, 7, 15)
gh

#> GOHits object with 6 hits and 2 metadata columns:
#>            from        to | ontology evidence
#>       <integer> <integer> | <factor> <factor>
#>   [1]         5        11 |       BP      IEA
#>   [2]         2        15 |       BP      IDA
#>   [3]         3         5 |       BP      IEA
#>   [4]         3         4 |       MF      IDA
#>   [5]         3         5 |       MF      IEA
#>   [6]         2        11 |       CC      IDA
#>   -------
#>   nLnode: 7 / nRnode: 15