Methods for GOHits Objects

This page documents the S4 generics and methods defined for objects inheriting of the GOHits class. The GOHits class directly extends Hits and thus inherits of all methods defined for the parent class. In the usage below, object represents an object of class inheriting from GOHits, and value is an object of a class specified in the S4 method signature or as outlined in 'Accessors'.

ontology(object)

ontology(object) <- value

evidence(object)

evidence(object) <- value

# S4 method for GOHits
evidence(object)

# S4 method for GOHits
evidence(object) <- value

# S4 method for GOHits
ontology(object)

# S4 method for GOHits
ontology(object) <- value

Arguments

object
An object of class inheriting from GOHits.

value
An object of a class specified in the S4 method signature or as outlined in 'Accessors'.

Accessors

evidence(object) returns a factor indicating the evidence code for each relation.

ontology(object) returns a factor indicating the ontology code for each relation.

See also

Examples

# Controlled vocabulary ----

GOEvidenceCodes

#>                                                     EXP 
#>                              "Inferred from Experiment" 
#>                                                     IDA 
#>                            "Inferred from Direct Assay" 
#>                                                     IPI 
#>                    "Inferred from Physical Interaction" 
#>                                                     IMP 
#>                        "Inferred from Mutant Phenotype" 
#>                                                     IGI 
#>                     "Inferred from Genetic Interaction" 
#>                                                     IEP 
#>                      "Inferred from Expression Pattern" 
#>                                                     HTP 
#>              "Inferred from High Throughput Experiment" 
#>                                                     HDA 
#>            "Inferred from High Throughput Direct Assay" 
#>                                                     HMP 
#>        "Inferred from High Throughput Mutant Phenotype" 
#>                                                     HGI 
#>     "Inferred from High Throughput Genetic Interaction" 
#>                                                     HEP 
#>      "Inferred from High Throughput Expression Pattern" 
#>                                                     IBA 
#>   "Inferred from Biological characteristic of Ancestor" 
#>                                                     IBD 
#> "Inferred from Biological characteristic of Descendant" 
#>                                                     IKR 
#>          "Inferred loss due to absence of Key Residues" 
#>                                                     IRD 
#>                  "Inferred loss after Rapid Divergence" 
#>                                                     ISS 
#>       "Inferred from Sequence or structural Similarity" 
#>                                                     ISO 
#>                      "Inferred from Sequence Orthology" 
#>                                                     ISA 
#>                      "Inferred from Sequence Alignment" 
#>                                                     ISM 
#>                          "Inferred from Sequence Model" 
#>                                                     IGC 
#>                         "Inferred from Genomic Context" 
#>                                                     RCA 
#>         "Inferred from Reviewed Computational Analysis" 
#>                                                     TAS 
#>                            "Traceable Author Statement" 
#>                                                     NAS 
#>                        "Non-traceable Author Statement" 
#>                                                      IC 
#>                                   "Inferred by Curator" 
#>                                                      ND 
#>                          "No biological Data available" 
#>                                                     IEA 
#>                   "Inferred from Electronic Annotation" 

GOOntologyCodes

#>                   BP                   MF                   CC 
#> "Biological Process" "Molecular Function" "Cellular Component" 

# Constructor ----

from <- c(5, 2, 3, 3, 3, 2)
to <- c(11, 15, 5, 4, 5, 11)
ontology <- factor(c("BP", "BP", "BP", "MF", "MF", "CC"))
evidence <- factor(c("IEA", "IDA", "IEA", "IDA", "IEA", "IDA"))

gh <- GOHits(from, to, evidence, ontology, 7, 15)
gh

#> GOHits object with 6 hits and 2 metadata columns:
#>            from        to | ontology evidence
#>       <integer> <integer> | <factor> <factor>
#>   [1]         5        11 |       BP      IEA
#>   [2]         2        15 |       BP      IDA
#>   [3]         3         5 |       BP      IEA
#>   [4]         3         4 |       MF      IDA
#>   [5]         3         5 |       MF      IEA
#>   [6]         2        11 |       CC      IDA
#>   -------
#>   nLnode: 7 / nRnode: 15

# Accessors ----

evidence(gh)

#> [1] IEA IDA IEA IDA IEA IDA
#> 26 Levels: EXP IDA IPI IMP IGI IEP HTP HDA HMP HGI HEP IBA IBD IKR IRD ... IEA

gh1 <- gh
evidence(gh1)[1] <- "HTP"
ontology(gh)

#> [1] BP BP BP MF MF CC
#> Levels: BP MF CC

gh1 <- gh
ontology(gh1)[1] <- "MF"