Methods for GOSets Objects

This page documents the S4 generics and methods defined for objects inheriting of the GOSets class. The GOSets class directly extends Sets and thus inherits of all methods defined for the parent class. In the usage below, object represents an object of class inheriting from GOSets, and value is an object of a class specified in the S4 method signature or as outlined in 'Accessors'.

# S4 method for GOSets
evidence(object)

# S4 method for GOSets
evidence(object) <- value

# S4 method for GOSets
ontology(object)

# S4 method for GOSets
ontology(object) <- value

Arguments

object
An object of class inheriting from GOSets.

value
An object of a class specified in the S4 method signature or as outlined in 'Accessors'.

Accessors

evidence(object) returns a factor indicating the evidence code for each relation.

ontology(object) returns a factor indicating the ontology code for each relation.

See also

Examples

# Constructor ----

# Fetch a sample of GO annotations
library(org.Hs.eg.db)
go_sets <- import(org.Hs.egGO)

#> 'select()' returned 1:1 mapping between keys and columns

#> Coercing evidence to factor

#> Coercing ontology to factor

relations <- as.data.frame(head(go_sets))

gs <- GOSets(relations)

# Accessors ----

evidence(gs)

#> [1] TAS ND  TAS IDA TAS TAS
#> 26 Levels: EXP IDA IPI IMP IGI IEP HTP HDA HMP HGI HEP IBA IBD IKR IRD ... IEA

gs1 <- gs
evidence(gs1)[1] <- "EXP"
ontology(gs)

#> [1] BP BP BP BP BP BP
#> Levels: BP MF CC

gs1 <- gs
ontology(gs1)[1] <- "CC"