IdVector Class

The IdVector class extends the Vector class to implement a container that hold a vector of character identifiers. Subclasses of IdVector may be defined to enable method dispatch according to the nature of the identifiers (e.g., ENTREZ gene, Gene Ontology term).

IdVector(ids = character(0))

EntrezIdVector(ids)

EnsemblIdVector(ids)

GOIdVector(ids)

Arguments

ids
character. Identifiers.

Value

An IdVector object.

Slots

ids

character. Identifiers.

See also

Examples

# Constructor ----

iv <- IdVector(ids=head(LETTERS, 6))
mcols(iv) <- DataFrame(row.names = ids(iv), field1=runif(length(iv)))
iv

#> IdVector of length 6 with 6 unique identifiers
#> Ids: A, B, C, D, ...
#> Metadata: field1 (1 column)

# Subsetting ----

iv[1:5]

#> IdVector of length 5 with 5 unique identifiers
#> Ids: A, B, C, D, ...
#> Metadata: field1 (1 column)

# Identifiers/Names ----

ids(iv)

#> [1] "A" "B" "C" "D" "E" "F"

names(iv)

#> [1] "A" "B" "C" "D" "E" "F"

# EntrezIdVector ----

library(org.Hs.eg.db)
eiv <- EntrezIdVector(keys(org.Hs.eg.db, keytype="ENTREZID"))
eiv

#> EntrezIdVector of length 61119 with 61119 unique identifiers
#> Ids: 1, 2, 3, 9, ...

# EnsemblIdVector ----

library(org.Hs.eg.db)
eiv <- EnsemblIdVector(keys(org.Hs.eg.db, keytype="ENSEMBL"))
eiv

#> EnsemblIdVector of length 30292 with 30292 unique identifiers
#> Ids: ENSG00000121410, ENSG00000175899, ENSG00000256069, ENSG00000171428, ...

# GOIdVector ----

library(org.Hs.eg.db)
giv <- GOIdVector(keys(org.Hs.eg.db, keytype = "GO"))
giv

#> GOIdVector of length 17983 with 17983 unique identifiers
#> Ids: GO:0002576, GO:0003674, GO:0005576, GO:0005615, ...