IdVector Class

The IdVector class extends the Vector class to implement a container that hold a vector of character identifiers. Subclasses of IdVector may be defined to enable method dispatch according to the nature of the identifiers (e.g., ENTREZ gene, Gene Ontology term).

IdVector(ids = character(0))

EntrezIdVector(ids)

Arguments

ids
character. Identifiers.

Value

An IdVector object.

Slots

ids

character. Unique identifiers.

See also

Examples

# Constructor ----

iv <- IdVector(ids=head(LETTERS, 6))
mcols(iv) <- DataFrame(row.names = ids(iv), field1=runif(length(iv)))
iv

#> IdVector of length 6 with 6 unique identifiers
#> Ids: A, B, C, D, ...
#> Metadata: field1 (1 column)

# Subsetting ----

iv[1:5]

#> IdVector of length 5 with 5 unique identifiers
#> Ids: A, B, C, D, ...
#> Metadata: field1 (1 column)

# Identifiers/Names ----

ids(iv)

#> [1] "A" "B" "C" "D" "E" "F"

names(iv)

#> [1] "A" "B" "C" "D" "E" "F"

# EntrezIdVector ----

library(org.Hs.eg.db)

#> Loading required package: AnnotationDbi

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: IRanges

#> 

ev <- EntrezIdVector(keys(org.Hs.eg.db))
# EntrezIdVector ----

library(org.Hs.eg.db)
ev <- EntrezIdVector(keys(org.Hs.eg.db))